+++ changelog.markdown
@@ -3,7 +3,13 @@
 * BREAKING CHANGE: Nuget dependencies are now encapsulate in a separate type. This allows to specify the upgrade strategy.
 * If the publishment of a Nuget package fails, then FAKE will try it again.
 * Added Changelog.markdown to FAKE deployment
- 
+
+## 1.52.0.0
+
+* Some smaller bugfixes
+* New dependency syntax with ==> and <=>
+* Tracing of StackTrace only if TargetHelper.PrintStackTraceOnError was set to true
+
 ## 1.50.0.0
 
 * New task DeleteDirs allows to delete multiple directories.